series_flint: use Flint C API instead of C++ API
