+++ symengine/mp_wrapper.h
@@ -4,11 +4,6 @@
 #include <symengine/symengine_rcp.h>
 #include <gmp.h>
 
-#if SYMENGINE_INTEGER_CLASS == SYMENGINE_FLINT
-#include <flint/fmpz.h>
-#include <flint/fmpq.h>
-#endif
-
 #define SYMENGINE_UI(f) f##_ui
 #define SYMENGINE_SI(f) f##_si
 
@@ -91,457 +86,14 @@
         return b op a;                                                         \
     }
 
-namespace SymEngine
-{
-
 #if SYMENGINE_INTEGER_CLASS == SYMENGINE_FLINT
 
-class fmpz_wrapper
-{
-private:
-    fmpz_t mp;
+#include <symengine/flint_wrapper.h>
 
-public:
-    template <typename T,
-              typename std::enable_if<std::is_integral<T>::value
-                                          && std::is_unsigned<T>::value,
-                                      int>::type
-              = 0>
-    inline fmpz_wrapper(const T i)
-    {
-        fmpz_init(mp);
-        fmpz_set_ui(mp, i);
-    }
-    template <typename T,
-              typename std::enable_if<std::is_integral<T>::value
-                                          && std::is_signed<T>::value,
-                                      int>::type
-              = 0>
-    inline fmpz_wrapper(const T i)
-    {
-        fmpz_init(mp);
-        fmpz_set_si(mp, i);
-    }
-    inline fmpz_wrapper()
-    {
-        fmpz_init(mp);
-    }
-    inline fmpz_wrapper(const mpz_t m)
-    {
-        fmpz_init(mp);
-        fmpz_set_mpz(mp, m);
-    }
-    inline fmpz_wrapper(const fmpz_t m)
-    {
-        fmpz_init(mp);
-        fmpz_set(mp, m);
-    }
-    inline fmpz_wrapper(const std::string &s, unsigned base = 10)
-    {
-        fmpz_init(mp);
-        fmpz_set_str(mp, s.c_str(), base);
-    }
-    inline fmpz_wrapper(const fmpz_wrapper &other)
-    {
-        fmpz_init(mp);
-        fmpz_set(mp, other.get_fmpz_t());
-    }
-    inline fmpz_wrapper(fmpz_wrapper &&other) SYMENGINE_NOEXCEPT
-    {
-        fmpz_init(mp);
-        fmpz_swap(mp, other.get_fmpz_t());
-    }
-    inline fmpz_wrapper &operator=(const fmpz_wrapper &other)
-    {
-        fmpz_set(mp, other.get_fmpz_t());
-        return *this;
-    }
-    inline fmpz_wrapper &operator=(fmpz_wrapper &&other) SYMENGINE_NOEXCEPT
-    {
-        fmpz_swap(mp, other.get_fmpz_t());
-        return *this;
-    }
-    inline ~fmpz_wrapper() SYMENGINE_NOEXCEPT
-    {
-        fmpz_clear(mp);
-    }
-    inline fmpz *get_fmpz_t()
-    {
-        return mp;
-    }
-    inline const fmpz *get_fmpz_t() const
-    {
-        return mp;
-    }
-    inline friend fmpz_wrapper operator+(const fmpz_wrapper &a,
-                                         const fmpz_wrapper &b)
-    {
-        fmpz_wrapper res;
-        fmpz_add(res.get_fmpz_t(), a.get_fmpz_t(), b.get_fmpz_t());
-        return res;
-    }
-    inline fmpz_wrapper operator+=(const fmpz_wrapper &a)
-    {
-        fmpz_add(mp, mp, a.get_fmpz_t());
-        return *this;
-    }
-    inline friend fmpz_wrapper operator-(const fmpz_wrapper &a,
-                                         const fmpz_wrapper &b)
-    {
-        fmpz_wrapper res;
-        fmpz_sub(res.get_fmpz_t(), a.get_fmpz_t(), b.get_fmpz_t());
-        return res;
-    }
-    inline fmpz_wrapper operator-=(const fmpz_wrapper &a)
-    {
-        fmpz_sub(mp, mp, a.get_fmpz_t());
-        return *this;
-    }
-    inline fmpz_wrapper operator-() const
-    {
-        fmpz_wrapper res;
-        fmpz_neg(res.get_fmpz_t(), mp);
-        return res;
-    }
-    inline friend fmpz_wrapper operator*(const fmpz_wrapper &a,
-                                         const fmpz_wrapper &b)
-    {
-        fmpz_wrapper res;
-        fmpz_mul(res.get_fmpz_t(), a.get_fmpz_t(), b.get_fmpz_t());
-        return res;
-    }
-    inline fmpz_wrapper operator*=(const fmpz_wrapper &a)
-    {
-        fmpz_mul(mp, mp, a.get_fmpz_t());
-        return *this;
-    }
-    inline friend fmpz_wrapper operator/(const fmpz_wrapper &a,
-                                         const fmpz_wrapper &b)
-    {
-        fmpz_wrapper res;
-        fmpz_tdiv_q(res.get_fmpz_t(), a.get_fmpz_t(), b.get_fmpz_t());
-        return res;
-    }
-    inline fmpz_wrapper operator/=(const fmpz_wrapper &a)
-    {
-        fmpz_tdiv_q(mp, mp, a.get_fmpz_t());
-        return *this;
-    }
-    inline friend fmpz_wrapper operator%(const fmpz_wrapper &a,
-                                         const fmpz_wrapper &b)
-    {
-        fmpz_wrapper res, tmp;
-        fmpz_tdiv_qr(tmp.get_fmpz_t(), res.get_fmpz_t(), a.get_fmpz_t(),
-                     b.get_fmpz_t());
-        return res;
-    }
-    inline fmpz_wrapper operator%=(const fmpz_wrapper &a)
-    {
-        fmpz_wrapper tmp;
-        fmpz_tdiv_qr(tmp.get_fmpz_t(), mp, mp, a.get_fmpz_t());
-        return *this;
-    }
-    inline fmpz_wrapper operator++()
-    {
-        fmpz_add_ui(mp, mp, 1);
-        return *this;
-    }
-    inline fmpz_wrapper operator++(int)
-    {
-        fmpz_wrapper orig = *this;
-        ++(*this);
-        return orig;
-    }
-    inline fmpz_wrapper operator--()
-    {
-        fmpz_sub_ui(mp, mp, 1);
-        return *this;
-    }
-    inline fmpz_wrapper operator--(int)
-    {
-        fmpz_wrapper orig = *this;
-        --(*this);
-        return orig;
-    }
-    inline friend bool operator==(const fmpz_wrapper &a, const fmpz_wrapper &b)
-    {
-        return fmpz_equal(a.get_fmpz_t(), b.get_fmpz_t()) == 1;
-    }
-    inline friend bool operator!=(const fmpz_wrapper &a, const fmpz_wrapper &b)
-    {
-        return fmpz_equal(a.get_fmpz_t(), b.get_fmpz_t()) != 1;
-    }
-    inline friend bool operator<(const fmpz_wrapper &a, const fmpz_wrapper &b)
-    {
-        return fmpz_cmp(a.get_fmpz_t(), b.get_fmpz_t()) < 0;
-    }
-    inline friend bool operator<=(const fmpz_wrapper &a, const fmpz_wrapper &b)
-    {
-        return fmpz_cmp(a.get_fmpz_t(), b.get_fmpz_t()) <= 0;
-    }
-    inline friend bool operator>(const fmpz_wrapper &a, const fmpz_wrapper &b)
-    {
-        return fmpz_cmp(a.get_fmpz_t(), b.get_fmpz_t()) > 0;
-    }
-    inline friend bool operator>=(const fmpz_wrapper &a, const fmpz_wrapper &b)
-    {
-        return fmpz_cmp(a.get_fmpz_t(), b.get_fmpz_t()) >= 0;
-    }
-    inline fmpz_wrapper operator<<=(unsigned long u)
-    {
-        fmpz_mul_2exp(mp, mp, u);
-        return *this;
-    }
-    inline fmpz_wrapper operator<<(unsigned long u) const
-    {
-        fmpz_wrapper res;
-        fmpz_mul_2exp(res.get_fmpz_t(), mp, u);
-        return res;
-    }
-    inline fmpz_wrapper operator>>=(unsigned long u)
-    {
-        fmpz_tdiv_q_2exp(mp, mp, u);
-        return *this;
-    }
-    inline fmpz_wrapper operator>>(unsigned long u) const
-    {
-        fmpz_wrapper res;
-        fmpz_tdiv_q_2exp(res.get_fmpz_t(), mp, u);
-        return res;
-    }
-};
-
-class mpz_view_flint
-{
-
-public:
-    mpz_view_flint(const fmpz_wrapper &i)
-    {
-        if (!COEFF_IS_MPZ(*i.get_fmpz_t())) {
-            mpz_init_set_si(m, *i.get_fmpz_t());
-        } else {
-            ptr = COEFF_TO_PTR(*i.get_fmpz_t());
-        }
-    }
-    operator mpz_srcptr() const
-    {
-        if (ptr == nullptr)
-            return m;
-        return ptr;
-    }
-    ~mpz_view_flint()
-    {
-        if (ptr == nullptr)
-            mpz_clear(m);
-    }
-
-private:
-    mpz_srcptr ptr = nullptr;
-    mpz_t m;
-};
-
-class fmpq_wrapper
-{
-private:
-    fmpq_t mp;
-
-public:
-    fmpq *get_fmpq_t()
-    {
-        return mp;
-    }
-    const fmpq *get_fmpq_t() const
-    {
-        return mp;
-    }
-    fmpq_wrapper()
-    {
-        fmpq_init(mp);
-    }
-    fmpq_wrapper(const mpz_t m)
-    {
-        fmpq_init(mp);
-        fmpz_set_mpz(fmpq_numref(mp), m);
-    }
-    fmpq_wrapper(const fmpz_t m)
-    {
-        fmpq_init(mp);
-        fmpz_set(fmpq_numref(mp), m);
-    }
-    fmpq_wrapper(const mpq_t m)
-    {
-        fmpq_init(mp);
-        fmpq_set_mpq(mp, m);
-    }
-    fmpq_wrapper(const fmpq_t m)
-    {
-        fmpq_init(mp);
-        fmpq_set(mp, m);
-    }
-    template <typename T,
-              typename std::enable_if<std::is_integral<T>::value
-                                          && std::is_unsigned<T>::value,
-                                      int>::type
-              = 0>
-    fmpq_wrapper(const T i)
-    {
-        fmpq_init(mp);
-        fmpz_set_ui(fmpq_numref(mp), i);
-    }
-    template <typename T,
-              typename std::enable_if<std::is_integral<T>::value
-                                          && std::is_signed<T>::value,
-                                      int>::type
-              = 0>
-    fmpq_wrapper(const T i)
-    {
-        fmpq_init(mp);
-        fmpz_set_si(fmpq_numref(mp), i);
-    }
-    fmpq_wrapper(const fmpz_wrapper &n, const fmpz_wrapper &d = 1)
-    {
-        fmpq_init(mp);
-        fmpz_set(fmpq_numref(mp), n.get_fmpz_t());
-        fmpz_set(fmpq_denref(mp), d.get_fmpz_t());
-    }
-    fmpq_wrapper(const fmpq_wrapper &other)
-    {
-        fmpq_init(mp);
-        fmpq_set(mp, other.get_fmpq_t());
-    }
-    fmpq_wrapper(fmpq_wrapper &&other)
-    {
-        fmpq_init(mp);
-        fmpq_swap(mp, other.get_fmpq_t());
-    }
-    fmpq_wrapper &operator=(const fmpq_wrapper &other)
-    {
-        fmpq_set(mp, other.get_fmpq_t());
-        return *this;
-    }
-    fmpq_wrapper &operator=(fmpq_wrapper &&other)
-    {
-        fmpq_swap(mp, other.get_fmpq_t());
-        return *this;
-    }
-    ~fmpq_wrapper()
-    {
-        fmpq_clear(mp);
-    }
-    const fmpz_wrapper &get_den() const
-    {
-        return reinterpret_cast<const fmpz_wrapper &>(*fmpq_denref(mp));
-    }
-    const fmpz_wrapper &get_num() const
-    {
-        return reinterpret_cast<const fmpz_wrapper &>(*fmpq_numref(mp));
-    }
-    fmpz_wrapper &get_den()
-    {
-        return reinterpret_cast<fmpz_wrapper &>(*fmpq_denref(mp));
-    }
-    fmpz_wrapper &get_num()
-    {
-        return reinterpret_cast<fmpz_wrapper &>(*fmpq_numref(mp));
-    }
-    friend fmpq_wrapper operator+(const fmpq_wrapper &a, const fmpq_wrapper &b)
-    {
-        fmpq_wrapper res;
-        fmpq_add(res.get_fmpq_t(), a.get_fmpq_t(), b.get_fmpq_t());
-        return res;
-    }
-    fmpq_wrapper operator+=(const fmpq_wrapper &a)
-    {
-        fmpq_add(mp, mp, a.get_fmpq_t());
-        return *this;
-    }
-    friend fmpq_wrapper operator-(const fmpq_wrapper &a, const fmpq_wrapper &b)
-    {
-        fmpq_wrapper res;
-        fmpq_sub(res.get_fmpq_t(), a.get_fmpq_t(), b.get_fmpq_t());
-        return res;
-    }
-    fmpq_wrapper operator-=(const fmpq_wrapper &a)
-    {
-        fmpq_sub(mp, mp, a.get_fmpq_t());
-        return *this;
-    }
-    fmpq_wrapper operator-() const
-    {
-        fmpq_wrapper res;
-        fmpq_neg(res.get_fmpq_t(), mp);
-        return res;
-    }
-    friend fmpq_wrapper operator*(const fmpq_wrapper &a, const fmpq_wrapper &b)
-    {
-        fmpq_wrapper res;
-        fmpq_mul(res.get_fmpq_t(), a.get_fmpq_t(), b.get_fmpq_t());
-        return res;
-    }
-    fmpq_wrapper operator*=(const fmpq_wrapper &a)
-    {
-        fmpq_mul(mp, mp, a.get_fmpq_t());
-        return *this;
-    }
-    friend fmpq_wrapper operator/(const fmpq_wrapper &a, const fmpq_wrapper &b)
-    {
-        fmpq_wrapper res;
-        fmpq_div(res.get_fmpq_t(), a.get_fmpq_t(), b.get_fmpq_t());
-        return res;
-    }
-    fmpq_wrapper operator/=(const fmpq_wrapper &a)
-    {
-        fmpq_div(mp, mp, a.get_fmpq_t());
-        return *this;
-    }
-    bool operator==(const fmpq_wrapper &other) const
-    {
-        return fmpq_equal(mp, other.get_fmpq_t());
-    }
-    bool operator!=(const fmpq_wrapper &other) const
-    {
-        return not(*this == other);
-    }
-    bool operator<(const fmpq_wrapper &other) const
-    {
-        return fmpq_cmp(mp, other.get_fmpq_t()) < 0;
-    }
-    bool operator<=(const fmpq_wrapper &other) const
-    {
-        return fmpq_cmp(mp, other.get_fmpq_t()) <= 0;
-    }
-    bool operator>(const fmpq_wrapper &other) const
-    {
-        return fmpq_cmp(mp, other.get_fmpq_t()) > 0;
-    }
-    bool operator>=(const fmpq_wrapper &other) const
-    {
-        return fmpq_cmp(mp, other.get_fmpq_t()) >= 0;
-    }
-};
-
-class mpq_view_flint
-{
-
-public:
-    mpq_view_flint(const fmpq_wrapper &i)
-    {
-        mpq_init(m);
-        fmpq_get_mpq(m, i.get_fmpq_t());
-    }
-    operator mpq_srcptr() const
-    {
-        return m;
-    }
-    ~mpq_view_flint()
-    {
-        mpq_clear(m);
-    }
-
-private:
-    mpq_t m;
-};
 #elif SYMENGINE_INTEGER_CLASS == SYMENGINE_GMP
+
+namespace SymEngine
+{
 
 class mpz_wrapper
 {
@@ -933,7 +485,12 @@ public:
     }
 };
 
+} // SymEngine
+
 #endif
+
+namespace SymEngine
+{
 
 #if SYMENGINE_INTEGER_CLASS == SYMENGINE_FLINT
 std::ostream &operator<<(std::ostream &os, const SymEngine::fmpq_wrapper &f);
