+++ symengine/series_flint.h
@@ -5,16 +5,15 @@
 #include <symengine/expression.h>
 
 #ifdef HAVE_SYMENGINE_FLINT
-#include <flint/flint.h>
-#include <flint/fmpq_polyxx.h>
+#include <symengine/flint_wrapper.h>
 
 namespace SymEngine
 {
 
-using fp_t = flint::fmpq_polyxx;
+using fp_t = fmpq_poly_wrapper;
 // Univariate Rational Coefficient Power SeriesBase using Flint
 class URatPSeriesFlint
-    : public SeriesBase<fp_t, flint::fmpqxx, URatPSeriesFlint>
+    : public SeriesBase<fp_t, fmpq_wrapper, URatPSeriesFlint>
 {
 public:
     URatPSeriesFlint(const fp_t p, const std::string varname,
@@ -29,19 +28,22 @@ public:
     static RCP<const URatPSeriesFlint>
     series(const RCP<const Basic> &t, const std::string &x, unsigned int prec);
     static fp_t var(const std::string &s);
-    static flint::fmpqxx convert(const Basic &x);
+    static fp_t convert(const rational_class &x);
+    static fp_t convert(const Rational &x);
+    static fp_t convert(const Integer &x);
+    static fp_t convert(const Basic &x);
     static inline fp_t mul(const fp_t &s, const fp_t &r, unsigned prec)
     {
-        return fp_t(flint::mullow(s, r, prec));
+        return s.mullow(r, prec);
     }
     static fp_t pow(const fp_t &s, int n, unsigned prec);
     static unsigned ldegree(const fp_t &s);
-    static inline flint::fmpqxx find_cf(const fp_t &s, const fp_t &var,
+    static inline fmpq_wrapper find_cf(const fp_t &s, const fp_t &var,
                                         unsigned deg)
     {
-        return flint::fmpqxx((s.get_coeff(deg)));
+        return s.get_coeff(deg);
     }
-    static flint::fmpqxx root(flint::fmpqxx &c, unsigned n);
+    static fmpq_wrapper root(fmpq_wrapper &c, unsigned n);
     static fp_t diff(const fp_t &s, const fp_t &var);
     static fp_t integrate(const fp_t &s, const fp_t &var);
     static fp_t subs(const fp_t &s, const fp_t &var, const fp_t &r,
@@ -51,70 +53,70 @@ public:
                                      unsigned int prec)
     {
         SYMENGINE_ASSERT(not s.get_coeff(0).is_zero());
-        return fp_t(flint::inv_series(s, prec));
+        return s.inv_series(prec);
     }
     static inline fp_t series_reverse(const fp_t &s, const fp_t &var,
                                       unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero()
                          and not s.get_coeff(1).is_zero());
-        return fp_t(flint::revert_series(s, prec));
+        return s.revert_series(prec);
     }
     static inline fp_t series_log(const fp_t &s, const fp_t &var,
                                   unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_one());
-        return fp_t(log_series(s, prec));
+        return s.log_series(prec);
     }
     static inline fp_t series_exp(const fp_t &s, const fp_t &var,
                                   unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(exp_series(s, prec));
+        return s.exp_series(prec);
     }
     static inline fp_t series_sin(const fp_t &s, const fp_t &var,
                                   unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(sin_series(s, prec));
+        return s.sin_series(prec);
     }
 
     static inline fp_t series_cos(const fp_t &s, const fp_t &var,
                                   unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(cos_series(s, prec));
+        return s.cos_series(prec);
     }
 
     static inline fp_t series_tan(const fp_t &s, const fp_t &var,
                                   unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(tan_series(s, prec));
+        return s.tan_series(prec);
     }
     static inline fp_t series_atan(const fp_t &s, const fp_t &var,
                                    unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(atan_series(s, prec));
+        return s.atan_series(prec);
     }
     static inline fp_t series_atanh(const fp_t &s, const fp_t &var,
                                     unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(atanh_series(s, prec));
+        return s.atanh_series(prec);
     }
     static inline fp_t series_asin(const fp_t &s, const fp_t &var,
                                    unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(asin_series(s, prec));
+        return s.asin_series(prec);
     }
     static inline fp_t series_asinh(const fp_t &s, const fp_t &var,
                                     unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(asinh_series(s, prec));
+        return s.asinh_series(prec);
     }
     static inline fp_t series_acos(const fp_t &s, const fp_t &var,
                                    unsigned int prec)
@@ -125,19 +127,19 @@ public:
                                    unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(sinh_series(s, prec));
+        return s.sinh_series(prec);
     }
     static inline fp_t series_cosh(const fp_t &s, const fp_t &var,
                                    unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(cosh_series(s, prec));
+        return s.cosh_series(prec);
     }
     static inline fp_t series_tanh(const fp_t &s, const fp_t &var,
                                    unsigned int prec)
     {
         SYMENGINE_ASSERT(s.get_coeff(0).is_zero());
-        return fp_t(tanh_series(s, prec));
+        return s.tanh_series(prec);
     }
     static inline fp_t series_lambertw(const fp_t &s, const fp_t &var,
                                        unsigned int prec)
@@ -178,14 +180,14 @@ public:
         if (ldeg != 0) {
             ss = s * pow(var, -ldeg, prec);
         }
-        flint::fmpqxx ct = find_cf(ss, var, 0);
+        fmpq_wrapper ct = find_cf(ss, var, 0);
         bool do_inv = false;
         if (n < 0) {
             n = -n;
             do_inv = true;
         }
 
-        flint::fmpqxx ctroot = root(ct, n);
+        fmpq_wrapper ctroot = root(ct, n);
         fp_t res_p = one, sn = fp_t(ss / ct);
         auto steps = step_list(prec);
         for (const auto step : steps) {
